+++ example/lib/line_chart/samples/line_chart_sample8.dart
@@ -1,5 +1,11 @@
+import 'dart:async';
+import 'dart:typed_data';
+import 'dart:ui' as ui;
+
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart' show rootBundle;
+import 'package:flutter_svg/flutter_svg.dart';
 
 class LineChartSample8 extends StatefulWidget {
   @override
@@ -14,26 +20,52 @@ class _LineChartSample8State extends State<LineChartSample8> {
 
   bool showAvg = false;
 
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      children: <Widget>[
-        AspectRatio(
-          aspectRatio: 1.70,
-          child: Container(
-            child: Padding(
-              padding: const EdgeInsets.only(right: 18.0, left: 12.0, top: 24, bottom: 12),
-              child: LineChart(
-                mainData(),
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
+  Future<ui.Image> loadImage(String asset) async {
+    ByteData data = await rootBundle.load(asset);
+    ui.Codec codec = await ui.instantiateImageCodec(data.buffer.asUint8List());
+    ui.FrameInfo fi = await codec.getNextFrame();
+    return fi.image;
   }
 
-  LineChartData mainData() {
+  Future<SizedPicture> loadSvg() async {
+    final String rawSvg =
+        '<svg height="14" width="14" xmlns="http://www.w3.org/2000/svg"><g fill="none" fill-rule="evenodd" transform="translate(-.000014)"><circle cx="7" cy="7" fill="#495DFF" r="7"/><path d="m7 10.9999976c1.6562389 0 2.99998569-1.34374678 2.99998569-2.99999283s-1.34374679-4.99998808-2.99998569-4.99998808c-1.6562532 0-3 3.34374203-3 4.99998808s1.3437468 2.99999283 3 2.99999283z" fill="#fff" fill-rule="nonzero"/></g></svg>';
+
+    final DrawableRoot svgRoot = await svg.fromSvgString(rawSvg, rawSvg);
+
+    final ui.Picture picture = svgRoot.toPicture();
+    final sizedPicture = SizedPicture(picture, 14, 14);
+    return sizedPicture;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return FutureBuilder<SizedPicture>(
+        future: loadSvg(),
+        builder: (BuildContext context, imageSnapshot) {
+          if (imageSnapshot.connectionState == ConnectionState.done) {
+            return Stack(
+              children: <Widget>[
+                AspectRatio(
+                  aspectRatio: 1.70,
+                  child: Container(
+                    child: Padding(
+                      padding: const EdgeInsets.only(right: 18.0, left: 12.0, top: 24, bottom: 12),
+                      child: LineChart(
+                        mainData(imageSnapshot.data),
+                      ),
+                    ),
+                  ),
+                ),
+              ],
+            );
+          } else {
+            return CircularProgressIndicator();
+          }
+        });
+  }
+
+  LineChartData mainData(SizedPicture sizedPicture) {
     return LineChartData(
       rangeAnnotations: RangeAnnotations(
         verticalRangeAnnotations: [
@@ -86,6 +118,11 @@ class _LineChartSample8State extends State<LineChartSample8> {
               labelResolver: (line) => 'V: ${line.x}',
             ),
           ),
+          VerticalLine(
+            x: 8.5,
+            color: Colors.transparent,
+            sizedPicture: sizedPicture,
+          )
         ],
       ),
       gridData: FlGridData(